+++ src/redux/effects/companyInfoEffects.js
@@ -1,0 +1,16 @@
+const companyInfoEffectCleanData = ({ payload }) => payload.map((company) => ({
+  image: company.image,
+  companyName: company.companyName,
+  ceo: company.ceo,
+  address: company.address,
+  zip: company.zip,
+  marketCap: company.mktCap,
+  ticker: company.symbol,
+  currency: company.currency,
+  StockPrice: company.price,
+  changes: company.changes,
+  description: company.description,
+  website: company.website,
+}));
+
+export default companyInfoEffectCleanData;
